Updated the package to work with version 1.0.3 of julia.
Since chains in IterTools was deprecated and flatten(xs) did not work
in the same fashion as the previous implementation, I decided to work
with the xs directly without flattening into a matrix. This ment that
logproblambdax became a ::Vector{Vector{Vector}} which ment that a lot
of operations that were previously written using broadcasting was
replaced by for loops in the multi-histogram functions.

Because of the additional writing it was decided to split out the part
of find_δlogprob that was involved in solving the non-linear equations
in a separate function to keep the code pr. function to a minimum.

In addition, a method of solving the FS equations that involved
logarithms was added as an optional choice if the exponentials are large
which they were in the test-cases I ran.

Single-histogram initial guess to the ΔlogZs was added which resulted in
appreciable speedup of convergence of the FS equations and even made the
non-logarithm method possible for solving the equations.

Because we wanted to find statistical errors of the reweighted points
using jackknifing it was necessary to solve FS equations multiple times.
To speedup the convergence we wanted to use the previous guess for ΔlogZs
for subsequent computations. Thus we introduced a new field in the
ReweightObj that made this possible. Hopefully this shouldn't take
up too much memory.